+++ repos/system_upgrade/el7toel8/actors/rpmscanner/actor.py
@@ -1,8 +1,6 @@
-import yum
-
 from leapp.actors import Actor
+from leapp.libraries.actors import get_package_repository_data
 from leapp.libraries.common.rpms import get_installed_rpms
-from leapp.libraries.stdlib import run
 from leapp.models import InstalledRPM, RPM
 from leapp.tags import IPUWorkflowTag, FactsPhaseTag
 
@@ -19,19 +17,9 @@ class RpmScanner(Actor):
     produces = (InstalledRPM,)
     tags = (IPUWorkflowTag, FactsPhaseTag)
 
-    def get_package_repository_data(self):
-        """ Return dictionary mapping package name with repository from which it was installed """
-        yum_base = yum.YumBase()
-        pkg_repos = {}
-        for pkg in yum_base.doPackageLists().installed:
-            pkg_repos[pkg.name] = pkg.ui_from_repo.lstrip('@')
-
-        return pkg_repos
-
-
     def process(self):
         output = get_installed_rpms()
-        pkg_repos = self.get_package_repository_data()
+        pkg_repos = get_package_repository_data()
 
         result = InstalledRPM()
         for entry in output:
